chore: pr fixes

+++ server/lib/routes/token.js
@@ -2,21 +2,26 @@
 
 const uniqid = require('uniqid')
 
+const validationSchema = {
+  body: {
+    type: 'object',
+    properties: {
+      subscriptionId: { type: 'string' },
+      secretId: { type: 'string' }
+    },
+    required: ['subscriptionId', 'secretId']
+  }
+}
+
 async function tokenRoutes(server) {
   server.route({
     method: 'PUT',
     url: '/api/token',
+    schema: validationSchema,
     preHandler: server.auth([server.authenticate]),
     handler: async (request, reply) => {
-      const { firebaseAdmin, log } = server
+      const { firebaseAdmin } = server
       const { subscriptionId = null, secretId = null } = request.body
-
-      if (!subscriptionId || !secretId) {
-        log.error(
-          `Invalid subscriptionId/secretId received for user-${request.user}`
-        )
-        return reply.code(400).send('Invalid request')
-      }
 
       const db = firebaseAdmin.firestore()
 
@@ -27,8 +32,7 @@ async function tokenRoutes(server) {
         .doc(secretId)
         .set({
           token,
-          subscriptionId,
-          userId: request.user
+          subscriptionId
         })
       reply.send({ token })
     }
@@ -44,12 +48,37 @@ async function tokenRoutes(server) {
 
       const db = firebaseAdmin.firestore()
 
+      const secretRef = await db
+        .collection('tokens')
+        .doc(secretId)
+        .get()
+
+      if (secretRef.empty) {
+        return reply.code(404).send('Secret not found')
+      }
+
+      const { subscriptionId } = secretRef.docs[0].data()
+
+      const subscriptionRef = await db
+        .collection('subscriptions')
+        .where(
+          firebaseAdmin.firestore.FieldPath.documentId(),
+          '==',
+          subscriptionId
+        )
+        .where('userId', '==', request.user)
+        .get()
+
+      if (subscriptionRef.empty) {
+        return reply.code(403).send('Not authorized')
+      }
+
       await db
         .collection('tokens')
         .doc(secretId)
         .delete()
 
-      return reply.code(204).send()
+      reply.code(204).send()
     }
   })
 }
